build: compile the menu overlay under MinGW; LONG for g_qPending
The dinput8 rule predates the imgui overlay: scripthook_ovl.cpp and the
Dear ImGui sources are C++, so they now compile with the C++ compiler
into build/ objects and link in beside the C sources (still compiled by
the CC driver). libstdc++ and winpthread are linked static through a
-Bstatic window - the DLL must not take a runtime DLL with it into the
game, and the post-build guard now names the offending import and also
catches libstdc++/libgcc_s. d3dcompiler_47 is linked explicitly (MSVC
gets it from a pragma the backend carries).

g_qPending is typed LONG like every interlocked variable beside it;
GCC 14+ type-checks the intrinsic and the int declaration no longer
builds.

+++ scripthook_physics.c
@@ -187,7 +187,7 @@ extern void ShSceneTick(void);
 static volatile uint64_t g_qFn = 0;
 static uint64_t g_qArg[6];
 static volatile uint64_t g_qRet = 0;
-static volatile int g_qPending = 0;
+static volatile LONG g_qPending = 0;
 static volatile int g_qDone = 0;
 static volatile int g_qFloat = 0;
 static volatile int g_qSix = 0;
